+++ ElvUI_LuckyoneUI/Core/Update.lua
@@ -6,6 +6,28 @@ function L1UI:UpdateLayout(layout)
 
 	if not E.db.movers then E.db.movers = {} end
 
+	E.global["general"]["fadeMapWhenMoving"] = false
+
+	E.db["datatexts"]["fontOutline"] = "OUTLINE"
+	E.db["unitframe"]["units"]["arena"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["arena"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["focus"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["focus"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["party"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["party"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["pet"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["pet"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["player"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["player"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["raid"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["raid"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["raid40"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["raid40"]["healPrediction"]["enable"] = true
+	E.db["unitframe"]["units"]["target"]["healPrediction"]["absorbStyle"] = "WRAPPED"
+	E.db["unitframe"]["units"]["target"]["healPrediction"]["enable"] = true
+
+	L1UI:SwapTags('luckyone')
+
 	if layout == 'dps' then
 
 	elseif layout == 'healer' then
@@ -13,6 +35,4 @@ function L1UI:UpdateLayout(layout)
 	end
 
 	E:StaggeredUpdateAll(nil, true)
-
-	ReloadUI()
 end