+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+### Version 1.23 [20-04-2021]
+**Changes:**  
+Absorb Style setup  
+Heal Prediction setup  
+Added font outline for DataTexts  
+No longer fade worldmap while moving  
+**New Additions:**  
+Config section "Tags"  
+Config section "Skins" (WIP)  
+Tag: luckyone:health:percent  
+• Displays health percentage with 0 decimals at full health  
+• 1 decimal below 100% health and 2 decimals below 10% health  
+**Misc:**  
+Updated Credits  
+Updated WeakAuras on Wago  
+• Mostly Paladin, Warlock, Shaman, DH, DK, Druid  
+___
 ### Version 1.22 [12-04-2021]
 **Misc:**  
 WeakAura updates